Run rubocop without bundler
Rubocop should be run without "bundle _${BUNDLE_VERSION}_ exec",
since rubocop is installed as a part of the "netsoft-danger" gem.

And "netsoft-danger" is not included in Gemfile, but installed
separately on CI, which leads to Gem::Exception when rubocop
is run with "bundle exec".

Error message:
rubocop is not currently included in the bundle.