fix(bin/netsoft-circle): fix NoMethodError in the 'merge' command.
The `.from_hash` method is returning an Array of `SimpleCov::Result`
objects rather than a single `SimpleCov::Result`. We end up appending
this to another array, leaving us with an array of arrays instead of an
array of `SimpleCov::Result` objects as was expected.

The fix switches to `.concat` instead `<<`, there by leaving us with
an array of `SimpleCov::Result` objects. I tried the command on both
a recent result set and a result set from the last successful build:
the results were entirely the same, with both result sets returning
an array and triggering the same error, suggesting a code change for
this problem.